Modified filename. Remove trends for velocity images only.
Ver. 6., 03-Jun-2024, TF - Modified filename to adhere to the SoLO
Metadata standard. Ensure that trends are only removed for velocity
images (provided that one or more of the remove_*_trend keywords are set). New keyword fit_trend, if set together with one or both
remove_*_trend, remove a linear fit of the velocity trend instead of
removing the mean of each row and/or column.