Temporary fix to handle L2 files that have incorrect PIXLISTS value
Temporary fix to prevent crash when reading faulty L2 files: due to a bug in oslo_fits the PIXLISTS keyword may contain entries that should not be present in L2 files. PIXLISTS should either not be present or contain SATPIXLIST and only SATPIXLIST. For HDUs without any saturated pixels, but with missing telemetry packets, the PIXLISTS keyword may still be present due to this bug, and it may contain e.g. LOSTPLNPIXLIST.